restructure the navigation components and layout
I'm adding a top navigation combined with the side one, in order to detach everything that is not navigation from the side nav. Additionally, I'm gonna move all the
user and logout stuff to the top navigation. The notifications are also gonna be here, which will make it much easier to position properly. Also, in further commits I
will enable resizing of the navigation to get more screen space. This commit also fixes the problematic scroll with a fixed view. Now the navigation is fixed, and the
page can just scroll.